Updates to page one along with page 2 and 3
Need several ad templates for page 2 of my issue. I am not sure if I just have a bad scan of page 3 or what the issue is, but I have many errors and blank spaces when I read the page. I am not sure if the OCR program normally creates this many errors or it is a bad scan.