Add files via upload
I have finished pages 1-4 and will upload later pages when they are completed. One of the tables on my page 4 was not available so i left a comment in place instead. I still need to change the dates in XML for page 1 because I edited most of that page in author mode.