Deliverable 5
Since everyday of my week has 8 pages, I’ve decided to try a different
approach. I have done all of the page 1’s of the week, and I am
currently working on page 2 and so on. I have found this to be a lot
easier because I am doing the same thing for every page and a lot of it
repeats which makes it go a lot faster vs. working on multiple things
at once.