Full Monday Issue 1906-01-15
This is the entirety of the Monday issue.  My issues have 8 pages and I am shooting for upmost accuracy, that is why it has been taking me a while.  My issues are tedious but I am handling it.  The page 3 of my issue is almost completely crooked, in it's natural state.  It would take hours on end to type the entire page.  I even tried using FineReader over Cisdem.