Add files via upload
Sorry for this being late, I plan to catch up with the work this week. I've been plagued with a midterm for physics that I had to pass, because I was not doing so hot. I also like my files to be mostly complete and accurate so I spent a lot of time working on the first day thinking it would help for the other ones and it did but not to the point I thought. Anyways I plan to be caught up with my work by before spring breaks starts, I hope understand and thank you.