Monday Issue (Final)
This is my final totally completed Monday issue. I used regex on people's name, i also did dateline, index numbers, and placeNames. I did not get all of them especially for people names. Those horrible financial tables are also changed. There were also some templates missing so those ads I did not put in the issue and I skipped one article thing when I was ocring because it had a table and I thought it was a financial table, but it wasn't. So i did not include it in this issue, because it would be to time consuming to write it out from scratch.