1906-12-03
It is definitely time to buckle down and get around to editing. A lot of things went wrong with getting this pull request together but by far what was most to blame was my work ethic.
Time to spend all day editing for... two days probably. Maybe that will at least get me in the habit.