finished
Thank you for the great semester :) The workload was honestly so bad but I really enjoyed having you to teach the class as you kept it interesting and was always there for help and seemed to actually care about us as students. It isn't often that teachers are this way in such a rigorous class and I really have to give you credit for teaching a class like this where you likely receive so much hate due to the workload. I wish you the very best in your future inquiries into the Egyptian Gazette as well as your other projects and hope that I will see you around sometime.

Cheers,
Tryfon